Use scripts in deno_automation
