+++ scripts/01_setup.ts
@@ -1,8 +1,8 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
-import { Repos } from "./helpers/mod.ts";
+import { Repos } from "./repos.ts";
 
-const repos = new Repos();
+const repos = await Repos.load();
 
 // Ensure repos are latest main
 for (const repo of repos.nonDenoAstRepos()) {
