+++ scripts/deps.ts
@@ -1,3 +1,3 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
-export * from "./repo.ts";
+export * from "../../../deno_automation/mod.ts";
