--- scripts/helpers/deps.ts
@@ -1,4 +1,0 @@
-// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
-
-export * as path from "https://deno.land/std@0.113.0/path/mod.ts";
-export * as semver from "https://deno.land/x/semver@v1.4.0/mod.ts";
